requirements.txt
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.1.2
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg-binary==3.1.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-environ==0.4.54
setuptools==65.5.0
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
